requirements.txt
tqdm
Pillow
transformers
diffusers>=0.3.0
scipy
jupyterlab
